requirements.txt
cachetools==4.2.2
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.3
google-api-core==1.31.0
google-auth==1.33.0
google-cloud-texttospeech==2.5.0
googleapis-common-protos==1.53.0
grpcio==1.38.1
idna==3.2
packaging==21.0
pdfminer.six==20200517
pdfplumber==0.5.28
Pillow==8.3.1
proto-plus==1.19.0
protobuf==3.17.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycryptodome==3.10.1
pyparsing==2.4.7
pytz==2021.1
requests==2.26.0
rsa==4.7.2
six==1.16.0
sortedcontainers==2.4.0
urllib3==1.26.6
Wand==0.6.6
